Update SingleImageDropzone to ImageDropzone and make improvements
