+++ components/ImageDropzone/ImageDropzone.module.css
@@ -1,0 +1,48 @@
+.single_dropzone {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    border: 1px solid #373737;
+    padding: 0;
+    position: relative;
+}
+
+.image_overlay {
+    background-color: #00000077;
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding-top: 20px;
+}
+
+.image_preview {
+    margin-bottom: 10px;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    max-height: 200px;
+}
+
+.uploaded_images {
+    display: flex;
+}
+
+.uploaded_image {
+    object-fit: cover;
+    margin: 1px;
+}
+
+.uploaded_image_wrapper {
+    margin: 4px;
+}
+
+.uploaded_image_wrapper:hover {
+    background-color: red;
+}
+
+.uploaded_image:hover {
+    opacity: 0.5;
+}
