--- components/SingleImageDropzone/SingleImageDropzone.module.css
@@ -1,14 +1,0 @@
-.single_dropzone {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    border: 1px solid #373737;
-    padding: 10px 0;
-}
-
-.image_preview {
-    margin-bottom: 10px;
-    width: 100%;
-    height: 100px;
-    object-fit: cover;
-}
